+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.9
+* Update for current Mapbender API conventions (no functional changes)
+
+## 1.0.8
+* Fix coordinate display rounding drift on map click
+* Fix coordinate display rounding drift on coordinate input changes
+* Fix error when trying to center on map after external srs change
+* Fix internal server error when submitting configuration form with validation error
+* Drop "type" configuration option; auto-determine popup vs sidepane operation mode from containing region
+
 ## 1.0.7.2
 * Resolve Symfony 3 incompatibilities
 * Fix errors when using "Coordinate search" button with multiple Coordinates Utility instances in the same application